Restructure project to frontend and backend
