+++ README.md
@@ -1,6 +1,6 @@
 # Dashboard Ambiental Costero
 
-Este repositorio contiene una demostración de un panel de monitoreo ambiental compuesto por un **frontend en React** y un **backend en Express**. Supabase se utiliza como base de datos y servicio de autenticación.
+Este repositorio contiene una demostración de un panel de monitoreo ambiental compuesto por un **frontend en React** y un **backend en Express**. Supabase se utiliza como base de datos y servicio de autenticación. El proyecto está dividido en dos carpetas: `frontend` y `backend`.
 
 ## Requisitos
 - Node.js 18 o superior
@@ -9,14 +9,14 @@ Este repositorio contiene una demostración de un panel de monitoreo ambiental c
 
 ## Instalación manual
 ### Frontend
-1. Copia el archivo `accesibilidad-demo/.env` y coloca tus datos de `REACT_APP_OPENWEATHER_KEY`, `REACT_APP_SUPABASE_URL` y `REACT_APP_SUPABASE_KEY`.
+1. Copia el archivo `frontend/.env` y coloca tus datos de `REACT_APP_OPENWEATHER_KEY`, `REACT_APP_SUPABASE_URL` y `REACT_APP_SUPABASE_KEY`.
 2. Instala las dependencias ejecutando:
    ```bash
-   npm install --force --prefix accesibilidad-demo
+   npm install --force --prefix frontend
    ```
 3. Inicia el servidor de desarrollo con:
    ```bash
-   npm start --prefix accesibilidad-demo
+   npm start --prefix frontend
    ```
    La aplicación estará disponible en `http://localhost:3000`.
 
@@ -45,7 +45,7 @@ En la carpeta `scripts` encontrarás `supabase_schema.sql` con el script para cr
 ## Pruebas
 Para ejecutar la suite de tests del frontend utiliza:
 ```bash
-CI=true npm test --silent --runInBand --prefix accesibilidad-demo
+CI=true npm test --silent --runInBand --prefix frontend
 ```
 
 ## Licencia
