+++ backend/.env.sample
@@ -1,0 +1,3 @@
+SUPABASE_URL=
+SUPABASE_KEY=
+PORT=4000
