+++ backend/index.js
@@ -1,0 +1,25 @@
+require('dotenv').config();
+const express = require('express');
+const cors = require('cors');
+const { createClient } = require('@supabase/supabase-js');
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_KEY);
+
+app.get('/status', async (_req, res) => {
+  try {
+    const { error } = await supabase.from('cities').select('id').limit(1);
+    if (error) throw error;
+    res.json({ status: 'ok' });
+  } catch (e) {
+    res.status(500).json({ status: 'error' });
+  }
+});
+
+const PORT = process.env.PORT || 4000;
+app.listen(PORT, () => {
+  console.log(`Backend running on port ${PORT}`);
+});
