+++ frontend/README.md
@@ -1,6 +1,6 @@
 # Dashboard Ambiental Costero
 
-Este proyecto es una demostración de un panel ambiental construido con React. La aplicación se organiza siguiendo principios de **Clean Architecture** donde la lógica de dominio y servicios se ubican en carpetas específicas y los componentes de interfaz son puramente presentacionales.
+Este proyecto es una demostración de un panel ambiental construido con React. La aplicación se organiza de forma simple dentro de la carpeta `frontend` sin seguir Clean Architecture.
 
 ## Accesibilidad
 
@@ -14,13 +14,13 @@ Se han incorporado buenas prácticas de accesibilidad para cumplir con las pauta
 ## Instalación y pruebas
 
 1. Copia el archivo `.env` y coloca tu clave en `REACT_APP_OPENWEATHER_KEY` para obtener datos reales.
-2. Ejecuta `npm install --force` dentro del directorio `accesibilidad-demo` para instalar las dependencias.
+2. Ejecuta `npm install --force` dentro del directorio `frontend` para instalar las dependencias.
 3. Para las pruebas automáticas usa `CI=true npm test --silent --runInBand`.
 
 ## Ejecución
 
 ```bash
-npm start --prefix accesibilidad-demo
+npm start --prefix frontend
 ```
 
 La aplicación se abrirá en `http://localhost:3000`.
