+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "accesibilidad-demo",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
