+++ frontend/src/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from './presentation/App';
+import App from './App';
 jest.mock('react-leaflet');
 
 test('renders welcome page', () => {
