+++ frontend/src/__mocks__/react-leaflet.js
