+++ frontend/src/components/Header.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../../Landing.css';
+import '../Landing.css';
 import { useAuth } from '../AuthContext';
 
 export default function Header() {
