+++ frontend/src/components/Welcome.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../../Landing.css";
+import "../Landing.css";
 import Header from "./Header";
 
 export default function Welcome() {
