+++ frontend/src/data/icons.js
