+++ frontend/src/data/mockWeather.js
