+++ frontend/src/services/applyAccessibilityEffects.js
